+++ src/main/java/com/labsynch/labseer/service/SearchFormService.java
@@ -16,7 +16,7 @@ public interface SearchFormService {
 			int maxResults, Float similarity, SearchType searchType,
 			String outputFormat) throws IOException, CmpdRegMolFormatException;
 
-	public List<Integer> findParentIds(String molStructure,
+	public List<Long> findParentIds(String molStructure,
 			Integer maxResults, Float similarity, SearchType searchType,
 			List<String> projects) throws IOException, CmpdRegMolFormatException;
